// frontend/src/types/permission.types.ts
export enum Permission {
  // USERS
  CREATE_USER = "CREATE_USER",
  UPDATE_USER = "UPDATE_USER",
  DELETE_USER = "DELETE_USER",
  VIEW_USERS = "VIEW_USERS",

  // COMPANIES
  CREATE_COMPANY = "CREATE_COMPANY",
  UPDATE_COMPANY = "UPDATE_COMPANY",
  DELETE_COMPANY = "DELETE_COMPANY",
  VIEW_COMPANIES = "VIEW_COMPANIES",

  // RECIPES
  CREATE_RECIPE = "CREATE_RECIPE",
  UPDATE_RECIPE = "UPDATE_RECIPE",
  DELETE_RECIPE = "DELETE_RECIPE",
  VIEW_RECIPES = "VIEW_RECIPES",

  // INGREDIENTS
  CREATE_INGREDIENT = "CREATE_INGREDIENT",
  UPDATE_INGREDIENT = "UPDATE_INGREDIENT",
  DELETE_INGREDIENT = "DELETE_INGREDIENT",
  VIEW_INGREDIENTS = "VIEW_INGREDIENTS",

  // AUDIT
  VIEW_AUDIT = "VIEW_AUDIT",

  // EVENTS
  CREATE_EVENT = "CREATE_EVENT",
  UPDATE_EVENT = "UPDATE_EVENT",
  VIEW_EVENTS = "VIEW_EVENTS",

  // QUOTATIONS
  CREATE_QUOTATION = "CREATE_QUOTATION",
  VIEW_QUOTATIONS = "VIEW_QUOTATIONS",
}